requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.2.5
hydra-core==1.3.2
msrest==0.7.1
omegaconf==2.3.0